Dev order

+++ src/main/java/com/uade/tpo/demo/service/OrderServiceImpl.java
@@ -1,27 +1,19 @@
 package com.uade.tpo.demo.service;
 
-import com.uade.tpo.demo.controllers.config.JwtService;
 import com.uade.tpo.demo.entity.Order;
+import com.uade.tpo.demo.entity.OrderStatus;
 import com.uade.tpo.demo.entity.User;
-import com.uade.tpo.demo.exceptions.OrderDuplicateException;
+import com.uade.tpo.demo.exceptions.OrderNotFoundException;
 import com.uade.tpo.demo.repository.OrderRepository;
 import com.uade.tpo.demo.repository.UserRepository;
-import jakarta.servlet.FilterChain;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.lang.NonNull;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
@@ -34,40 +26,50 @@ public class OrderServiceImpl implements OrderService {
     @Autowired
     private UserRepository userRepository;
 
+    @Transactional
     public Page<Order> getOrders(PageRequest pageable) {
         return orderRepository.findAll(pageable);
     }
 
-    public Optional<Order> getOrderById(Long orderId) {
-        return orderRepository.findById(orderId);
+    @Transactional
+    public Order getOrderById(Long orderId) throws OrderNotFoundException{
+        Optional<Order> order = orderRepository.findById(orderId);
+        if (order.isPresent()) {
+            return order.get();
+        }
+        throw new OrderNotFoundException();
     }
 
 
-    public Order createOrder(Double total) throws OrderDuplicateException {
+    @Transactional
+    public Order createOrder(Double total, OrderStatus status) {
 
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         String email = auth.getName();
         User user = userRepository.findUserByEmail(email);
 
-        return orderRepository.save(new Order(total, user));
+        return orderRepository.save(new Order(total, user, status));
     }
 
-
-
-
-
-    @Override
-    public Order updateOrder(Long orderId, String status, Double total) {
-        return null;
+    @Transactional
+    public Order updateOrder(Long orderId, OrderStatus status, Double total) throws OrderNotFoundException {
+        Optional<Order> order = orderRepository.findById(orderId);
+        if (order.isPresent()) {
+            Order orderToUpdate = order.get();
+            orderToUpdate.setStatus(status);
+            orderToUpdate.setTotal(total);
+            return orderRepository.save(orderToUpdate);
+        }
+        throw new OrderNotFoundException();
     }
 
-    @Override
-    public Page<Order> getOrdersByUserId(Long userId, PageRequest pageRequest) {
-        return null;
-    }
-
-    @Override
-    public void deleteOrder(Long orderId) {
-        orderRepository.deleteById(orderId);
+    @Transactional
+    public Order deleteOrder(Long orderId) throws OrderNotFoundException{
+        Optional<Order> order = orderRepository.findById(orderId);
+        if (order.isPresent()) {
+            orderRepository.deleteById(orderId);
+            return order.get();
+        }
+        throw new OrderNotFoundException();
     }
 }
